+++ wrapper/configs/__init__.py
@@ -1,2 +1,2 @@
-from .argument_config import ArgumentConfig
-from .directory_config import DirectoryConfig
+from .arguments import ArgumentConfig
+from .directories import DirectoryConfig
